requirements.txt
flask
flask-mail
python-dotenv
cloudinary
supabase
gunicorn
requests
postgrest